Added code for WM8741/2 control over SPI.

+++ Core/Inc/pcm9211.h
@@ -86,8 +86,8 @@
 
 void PCM9211_Initialize(SPI_HandleTypeDef* h);
 void PCM9211_Setup(void);
-void PCM9211_Write(const uint32_t reg, const uint8_t data);
-uint8_t PCM9211_Read(const uint32_t reg);
-uint32_t PCM9211_WriteCheck(const uint32_t reg, const uint8_t data);
+void PCM9211_Write(const uint8_t reg, const uint8_t data);
+uint8_t PCM9211_Read(const uint8_t reg);
+uint32_t PCM9211_WriteCheck(const uint8_t reg, const uint8_t data);
 
 #endif /* INC_PCM9211_H_ */
